requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
Pillow==10.2.0
opencv-python==4.9.0.80 